Initial architecture + move palette logic into new file
